requirements.txt
Flask==3.0.3
flask-cors==5.0.0
mysql-connector-python==9.0.0
qrcode==8.0
Pillow==10.4.0
gunicorn==21.2.0
python-dotenv==1.0.0
